Add support for pages in addition to files
Pages are similar to files in most ways, but are not currently
processed by Doxylink. This is very basic support, but at least for
statically included pages it will allow linking using the page name.

For example, if you set USE_MDFILE_AS_MAINPAGE to a Markdown file, you
will now be able to reference that file with:

:myapi:`index`

Similarly, named pages can be referenced just as named files can.